Added static dns for origin

+++ Backend/app/routers/auth.py
@@ -61,8 +61,15 @@ async def whoami(session: SessionDep, cookies: Annotated[Cookies, Cookie()]) -> 
     Returns:
         dict: A dictionary with the key "display_name" containing the user's full name.
     """
+    if not cookies.id_token:
+        raise HTTPException(status_code=401, detail="Token not provided")
+
     token_record = session.exec(select(TokenStore).where(
         TokenStore.token == cookies.id_token)).first()
+
+    if not token_record:
+        raise HTTPException(status_code=401, detail="Token is invalid or expired")
+
     db_user = session.exec(select(User).where(
         User.login == token_record.login)).first()
     return {"display_name": db_user.first_name + " " + db_user.last_name}
@@ -171,15 +178,33 @@ async def login(user: UserLoginRequest, session: SessionDep, response: Response)
     ))
     session.commit()
 
-    response.set_cookie(key="DxpAccessToken", value=token)
+    response.set_cookie(
+        key="DxpAccessToken",
+        value=token,
+        httponly=True,   
+        secure=True,     
+        samesite="None",  
+        max_age=3600,    
+        expires=(datetime.now() + timedelta(hours=1)).isoformat(),  
+    )
     return {"display_name": db_user.first_name + " " + db_user.last_name}
 
 
-@router.get("/logout", status_code=200, dependencies=[Depends(check_token)])
+@router.post("/logout", status_code=200, dependencies=[Depends(check_token)])
 async def logout(cookies: Annotated[Cookies, Cookie()], response: Response, session: SessionDep):
     token_record = session.exec(select(TokenStore).where(
         TokenStore.token == cookies.id_token)).first()
-    session.delete(token_record)
-    session.commit()
-    response.set_cookie(key="DxpAccessToken", value="")
+    if token_record:
+        session.delete(token_record)
+        session.commit()
+
+    response.set_cookie(
+        key="DxpAccessToken",
+        value="",
+        httponly=True,
+        secure=True,
+        samesite="None",
+        max_age=0,
+        expires="Thu, 01 Jan 1970 00:00:00 GMT",
+    )
     return "logout"
